backend-python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
docling==2.52.0
requests==2.32.4
python-dotenv==1.1.1
